--- src/components/browse/index.js
@@ -1,2 +1,0 @@
-import $ from './BrowseBooks.js.js.js';
-export default $;
