+++ src/navigation/NavBar.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { withRouter } from 'react-router-dom';
 import NavLinksNonAuth from './NavLinksNonAuth';
-import styles from '../../scss/components/NavBar.module.scss';
+import styles from './styles/NavBar.module.scss';
 
 // Currently NavBar only renders NavNonAuth
 // If we implement authentication, we can create a NavAuth with different links and styling
@@ -10,8 +10,8 @@ import styles from '../../scss/components/NavBar.module.scss';
 const NavBar = () => (
   <div className={styles.navbar}>
     <div className={styles.navBarContents}>
-    <h1 className={styles.appName}>Chatwise</h1>
-    <NavLinksNonAuth />
+      <h1 className={styles.appName}>Chatwise</h1>
+      <NavLinksNonAuth />
     </div>
   </div>
   // <div>{props.authUser ? <NavAuth /> : <NavNonAuth />}</div>
