+++ src/navigation/styles/NavBar.module.scss
@@ -1,6 +1,4 @@
-@import '../main.scss';
-
-@import '../main.scss';
+@import 'src/styles/scss/main.scss';
 
 .navbar {
 	width: 100%;
