+++ src/navigation/styles/NavLinksNonAuth.module.scss
@@ -1,4 +1,4 @@
-@import '../main.scss';
+@import 'src/styles/scss/main.scss';
 
 .navLinksNonAuth {
 	width: 50%;
