+++ src/styles/scss/mixins/background_pattern.scss
